speech-to-text/websocket/jarvis/python/requirements.txt
pyaudio>=0.2.14
websockets>=12.0
python-dotenv>=1.0.0
groq>=0.4.0

